Moved Voc.Models to prevent import circles
Was an issue when using Vocabulary.parse_raw()
