+++ filip/semantics/vocabulary_configurator.py
@@ -11,7 +11,6 @@ from typing import List, Optional, Dict, Tuple, Set
 
 import requests
 import wget
-from pydantic import BaseModel, Field
 
 from filip.semantics.ontology_parser.post_processer import PostProcessor
 from filip.semantics.ontology_parser.rdfparser import RdfParser
@@ -25,7 +24,9 @@ from filip.semantics.vocabulary import \
     CombinedRelation, \
     DataFieldType, \
     DependencyStatement, \
-    VocabularySettings
+    VocabularySettings, \
+    LabelSummary
+
 
 # Blacklist containing all labels that are forbidden for entities to have
 label_blacklist = list(keyword.kwlist)
@@ -39,84 +40,6 @@ label_blacklist.extend(["str", "int", "float", "complex", "list", "tuple",
 
 # Whitelist containing all chars that an entity label can consist of
 label_char_whitelist = ascii_letters + digits + "_"
-
-
-class LabelSummary(BaseModel):
-    """
-    Model holding all information for label conflicts in a vocabulary
-    """
-    class_label_duplicates: Dict[str, List[Entity]] = Field(
-        description="All Labels that are used more than once for class_names "
-                    "on export."
-                    "Key: Label, Values: List of entities with key label"
-    )
-    field_label_duplicates: Dict[str, List[Entity]] = Field(
-        description="All Labels that are used more than once for property_names"
-                    "on export."
-                    "Key: Label, Values: List of entities with key label"
-    )
-    datatype_label_duplicates: Dict[str, List[Entity]] = Field(
-        description="All Labels that are used more than once for datatype "
-                    "on export." 
-                    "Key: Label, Values: List of entities with key label"
-    )
-
-    blacklisted_labels: List[Tuple[str, Entity]] = Field(
-        description="All Labels that are blacklisted, "
-                    "Tuple(Label, Entity with label)"
-    )
-    labels_with_illegal_chars: List[Tuple[str, Entity]] = Field(
-        description="All Labels that contain illegal characters, "
-                    "Tuple(Label, Entity with label)"
-    )
-
-    def is_valid(self) -> bool:
-        """Test if Labels are valid
-
-        Returns:
-            bool, True if no entries exist
-        """
-        return len(self.class_label_duplicates) == 0 and \
-               len(self.field_label_duplicates) == 0 and \
-               len(self.datatype_label_duplicates) == 0 and \
-               len(self.blacklisted_labels) == 0 and \
-               len(self.labels_with_illegal_chars) == 0
-
-    def __str__(self):
-        res = ""
-
-        def print_collection(collection):
-            sub_res = ""
-            for key, values in collection.items():
-                sub_res += f"\t{key}: "
-                for v in values:
-                    sub_res += f" \n\t\t{v.iri}"
-                sub_res += "\n"
-
-            if len(collection) == 0:
-                sub_res += "\t/\n"
-            return sub_res
-
-        def print_list(collection):
-            sub_res = ""
-            for key, value in collection:
-                sub_res += f"\t{key}: \t {value.iri}"
-                sub_res += "\n"
-            if len(collection) == 0:
-                sub_res += "\t/\n"
-            return sub_res
-
-        res += "class_label_duplicates:\n"
-        res += print_collection(self.class_label_duplicates)
-        res += "field_label_duplicates:\n"
-        res += print_collection(self.field_label_duplicates)
-        res += "datatype_label_duplicates:\n"
-        res += print_collection(self.datatype_label_duplicates)
-        res += "blacklisted_labels:\n"
-        res += print_list(self.blacklisted_labels)
-        res += "labels_with_illegal_chars:\n"
-        res += print_list(self.labels_with_illegal_chars)
-        return res
 
 
 class VocabularyConfigurator:
